generating SMS actions, saving to DB
