+++ src/server/controllers/principles.js
@@ -5,7 +5,6 @@ const mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
 const Principle = mongoose.model('Principle');
 const { validationResult } = require('express-validator');
-const isEmpty = require('../utils/validate').isEmpty;
 
 exports.add = async (req, res, next) => {
   try {
@@ -41,13 +40,8 @@ exports.addMany = async (req, res, next) => {
   try {
     const uid = req.body.uid;
     const content = req.body.content;
-    if (isEmpty(content)) {
-      throw new Error('expected non empty value for content');
-    }
+
     const principlesArray = content.map((entry, key) => {
-      if (isEmpty(entry.content)) {
-        throw new Error('expected non empty value for content');
-      }
       return new Principle({
         content: entry.content,
         owner: uid
@@ -98,13 +92,9 @@ exports.delete = async (req, res, next) => {
     if (!errors.isEmpty()) {
       throw new Error(JSON.stringify({ errors: errors.array() }));
     }
-
     const uid = req.body.uid;
     const targetId = req.body.id;
 
-    if (isEmpty(targetId)) {
-      throw new Error('expected non empty value for deletion id');
-    }
     // Id in the request is the ObjectId of principle to delete
     const query = {
       _id: targetId,
@@ -153,10 +143,6 @@ exports.update = async (req, res, next) => {
     const uid = req.body.uid;
     const targetId = req.body.id;
     const content = req.body.content;
-
-    if (isEmpty(targetId) || isEmpty(content)) {
-      throw new Error('expected non empty value for update id and content');
-    }
     const query = {
       _id: targetId,
       owner: uid
