+++ src/server/server.js
@@ -10,20 +10,18 @@ const cookieParser = require('cookie-parser');
 
 const scheduler = require('./utils/scheduler');
 scheduler(); // will add SMS actions to DB
+
 // App
 const app = express();
-
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
   extended: true
 }));
 app.use(cors());
 app.use(cookieParser());
-
 app.get('/', (req, res) => {
   res.send('Hello from the server! :)\n');
 });
-
 app.use(routes);
 
 module.exports = app;
