--- src/server/utils/validate.js
@@ -1,7 +1,0 @@
-exports.isEmpty = (value) => {
-  if (value === undefined || value.length === 0) {
-    return true;
-  } else {
-    return false;
-  }
-};
